services/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
PyJWT==2.8.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
flask-swagger-ui==4.11.1
markupsafe==2.1.3
marshmallow==3.20.1